+++ Radio/index.js
@@ -1,265 +1,176 @@
-import React, { Component, PropTypes } from 'react'
-import compose from 'ramda/src/compose'
-import deepMerge from 'deepmerge'
-import {
-  notifyOnLowFPS,
-  themeable,
-  uncontrolled,
-  uniqueName
-} from '@klarna/higher-order-components'
+import React, {Component} from 'react'
 import {Motion, spring} from 'react-motion'
 import componentQueries from 'react-component-queries'
-import Option from './Option'
-import Collapsible from '../Collapsible'
-import defaultStyles from './styles'
-import getActiveElement from '../lib/getActiveElement'
-import ExpandLabel from './ExpandLabel'
+import {
+  uncontrolled,
+  uniqueName,
+  withDisplayName
+} from '@klarna/higher-order-components'
+import compose from 'ramda/src/compose'
+import Option, {OPTION_HEIGHT} from './Option'
+import Shadow from './Shadow'
+import debounce from '../lib/debounce'
+
+import * as palette from '../settings/palette'
 import * as breakpoints from '../settings/breakpoints'
 
-// The minimum/initial height of the ExpandLabel is 49 pixels. Unfortunately
-// it seems to be no way to get rid of this magic number without triggering
-// an animation on load
-const EXPAND_LABEL_INITIAL_HEIGHT = 49
+const SHADOW_HEIGHT = 20
+
+const styles = {
+  base: {
+    main: {
+      marginTop: -SHADOW_HEIGHT,
+      overflow: 'hidden',
+      paddingTop: SHADOW_HEIGHT,
+      position: 'relative'
+    },
+    optionWrapper: {
+      left: 0,
+      position: 'absolute',
+      width: '100%'
+    },
+    rogueElement: {
+      backgroundColor: palette.WHITE,
+      left: 0,
+      position: 'absolute',
+      width: '100%'
+    }
+  }
+}
+
+const update = (component) => {
+  component.setState({
+    optionContentSizes: Array.from(component.domElement.childNodes)
+      .slice(0, -1)
+      .map(domNode =>
+        domNode.getBoundingClientRect().height - OPTION_HEIGHT
+      )
+  })
+}
 
 class Radio extends Component {
   constructor () {
     super()
 
+    this.resizeListener = debounce(() => update(this))
+
     this.state = {
-      expandLabelInitialHeight: EXPAND_LABEL_INITIAL_HEIGHT
+      optionContentSizes: []
     }
   }
 
   componentDidMount () {
-    this.applyFocus()
+    window.addEventListener('resize', this.resizeListener)
+
+    this.resizeListener()
   }
 
-  componentDidUpdate () {
-    this.applyFocus()
-  }
-
-  applyFocus () {
-    if (
-      this.props.focus &&
-      this.refs[this.props.focus] &&
-      getActiveElement(document) !== this.refs[this.props.focus]
-    ) {
-      this.refs[this.props.focus].focus()
-    }
-  }
-
-  componentWillReceiveProps (nextProps) {
-    const { options, visibleOptions, value, onExpand, fullyExpanded } = nextProps
-
-    const shouldExpand = value &&
-      this.getOptionsLists(options, visibleOptions).collapsed.find(({key}) => key === value) != null
-
-    if (shouldExpand && !fullyExpanded) {
-      onExpand()
-    }
-  }
-
-  getOptionsLists (options, visibleOptions) {
-    return {
-      visible: typeof visibleOptions !== 'undefined'
-        ? options.slice(0, visibleOptions)
-        : options,
-      collapsed: typeof visibleOptions !== 'undefined'
-        ? options.slice(visibleOptions)
-        : []
-    }
+  componentWillUnmount () {
+    window.removeEventListener('resize', this.resizeListener)
   }
 
   render () {
     const {
-      borderless,
-      customize,
-      focus, // eslint-disable-line no-unused-vars
-      options,
-      disabled: allDisabled,
-      visibleOptions,
-      expandLabel,
-      fullyExpanded,
       name,
-      onBlur,
       onChange,
-      onExpand,
-      onFocus,
-      onEndFPSCollection,
-      onStartFPSCollection,
-      lowFPS,
+      options,
       padded,
-      styles,
+      radioMarkRight,
       value,
-      ...remainingProps
+      ...props
     } = this.props
+    const {optionContentSizes} = this.state
 
-    const finalStyles = deepMerge(defaultStyles, styles.radio)
-    const singleOption = options.length === 1
-    const baseStyle = customize ? { borderRadius: customize.borderRadius } : undefined
-    const labelStyle = customize ? { color: customize.textPrimaryColor } : undefined
-    const descriptionStyle = customize ? { color: customize.textSecondaryColor } : undefined
+    const selectedIndex = value != null
+      ? options.findIndex(({name}) => name === value)
+      : undefined
 
-    const optionLists = this.getOptionsLists(options, visibleOptions)
+    const rogueElementTranslateY = options.length * OPTION_HEIGHT +
+      (selectedIndex !== undefined
+        ? optionContentSizes[selectedIndex]
+        : 0)
 
-    const isExpanded = fullyExpanded || optionLists.collapsed.length === 0
-
-    const OptionWithProps = Option({
-      allDisabled,
-      onBlur,
-      value,
-      options,
-      borderless,
-      singleOption,
-      customize,
-      lowFPS,
-      labelStyle,
-      descriptionStyle,
-      onStartFPSCollection,
-      onEndFPSCollection,
-      onFocus,
-      onChange,
-      name,
-      padded,
-      styles
-    })
-
-    return (
-      <div
-        id={name}
-        style={{
-          ...finalStyles.base.main,
-          ...(borderless ? finalStyles.borderless.main : {}),
-          ...baseStyle
-        }}
-        {...remainingProps}>
-        {optionLists.visible.map(OptionWithProps)}
-
-        {optionLists.collapsed.length > 0 && <Collapsible
-          onStartFPSCollection={onStartFPSCollection}
-          onEndFPSCollection={onEndFPSCollection}
-          lowFPS={lowFPS}
-          minimumHeight={this.state.expandLabelInitialHeight}
-          collapsed={!isExpanded}>
-          <div>
-            {optionLists.collapsed.map(OptionWithProps)}
-          </div>
-        </Collapsible>}
-
-        {expandLabel && <Motion
+    return <div
+      ref={domElement => { this.domElement = domElement }}
+      style={{
+        ...styles.base.main,
+        height: (options.length * OPTION_HEIGHT) + SHADOW_HEIGHT +
+          (selectedIndex !== undefined
+            ? optionContentSizes[selectedIndex]
+            : 0)
+      }}
+      {...props}>
+      {options.map((optionProps, index) => {
+        const {content, name: optionName, ...option} = optionProps
+        return <Motion
+          key={optionName}
           style={{
-            opacity: spring(isExpanded ? 0 : 1),
-            height: spring(isExpanded ? 0 : this.state.expandLabelInitialHeight, {stiffness: 40, damping: 15})
+            translateY: spring(
+              (index * OPTION_HEIGHT) +
+              (selectedIndex !== undefined && selectedIndex < index
+                ? optionContentSizes[selectedIndex]
+                : 0)
+            )
           }}>
-          {({opacity, height}) => opacity > 0 && <div>
-            <div
-              style={{
-                position: 'absolute',
-                bottom: 0,
-                width: '100%',
-                backgroundColor: 'white',
-                boxShadow: 'white 0 0 5px',
-                height
-              }}
-            />
-            <ExpandLabel
-              onDOMElement={footer => footer && footer.getBoundingClientRect && (height => this.state.expandLabelInitialHeight !== height && this.setState({
-                expandLabelInitialHeight: height
-              }))(footer.getBoundingClientRect().height)}
-              onClick={onExpand}
-              label={expandLabel}
-              style={{opacity}}
-              styles={styles.expandLabel}
-            />
+          {({translateY}) => <div
+            style={{
+              ...styles.base.optionWrapper,
+              transform: `translateY(${translateY}px)`
+            }}>
+            <Option
+              index={index}
+              name={name}
+              id={optionName}
+              onChange={() => onChange(optionName)}
+              padded={padded}
+              previousSelected={index > 0
+                ? options[index - 1].name === value
+                : false
+              }
+              radioMarkRight={radioMarkRight}
+              selected={optionName === value}
+              {...option}>
+              {content}
+            </Option>
           </div>}
-        </Motion>}
-      </div>
-    )
-  }
-}
-
-Radio.propTypes = {
-  borderless: PropTypes.bool,
-  className: PropTypes.string,
-  customize: PropTypes.shape({
-    backgroundColor: PropTypes.string.isRequired,
-    borderRadius: PropTypes.string.isRequired,
-    bulletColor: PropTypes.string.isRequired,
-    textPrimaryColor: PropTypes.string.isRequired,
-    textSecondaryColor: PropTypes.string.isRequired
-  }),
-  disabled: PropTypes.bool,
-  visibleOptions: PropTypes.number,
-  expandLabel: PropTypes.string,
-  fullyExpanded: PropTypes.bool,
-  onExpand: PropTypes.func,
-  focus: PropTypes.string,
-  id: PropTypes.string,
-  name: PropTypes.string,
-  onBlur: PropTypes.func,
-  onChange: PropTypes.func,
-  onFocus: PropTypes.func,
-  onEndFPSCollection: PropTypes.func,
-  onStartFPSCollection: PropTypes.func,
-  lowFPS: PropTypes.bool,
-  options: PropTypes.array.isRequired,
-  padded: PropTypes.bool,
-  styles: PropTypes.shape({
-    radio: PropTypes.object,
-    option: PropTypes.object,
-    expandLabel: PropTypes.object
-  }),
-  value: PropTypes.any
-}
-
-Radio.defaultProps = {
-  fullyExpanded: false,
-  styles: {
-    radio: {},
-    option: {},
-    expandLabel: {},
-    radioMark: {}
+        </Motion>
+      })}
+      <Motion
+        style={{
+          translateY: spring(rogueElementTranslateY)
+        }}>
+        {({translateY}) => <div
+          style={{
+            ...styles.base.rogueElement,
+            height: SHADOW_HEIGHT +
+              (selectedIndex !== undefined
+                ? optionContentSizes[selectedIndex]
+                : 0),
+            transform: `translateY(${translateY}px)`
+          }}>
+          <Shadow
+            reversed
+            show={options[options.length - 1].name === value}
+          />
+        </div>}
+      </Motion>
+    </div>
   }
 }
 
 export default compose(
+  withDisplayName('Radio'),
   componentQueries(
     ({width}) => width > breakpoints.MOBILE_MAX_WIDTH
       ? {padded: true}
       : {padded: false}
   ),
-  notifyOnLowFPS({threshold: 10}),
-  uncontrolled({
-    prop: 'fullyExpanded',
-    defaultProp: 'autoFullyExpanded',
-    handlers: {
-      onExpand: () => () => true
-    }
-  }),
-  uncontrolled({
-    prop: 'focus',
-    defaultProp: 'autoFocus',
-    handlers: {
-      onFocus: () => field => field,
-      onBlur: () => () => undefined
-    }
-  }),
   uncontrolled({
     prop: 'value',
     defaultProp: 'defaultValue',
     handlers: {
-      onChange: () => value => value
+      onChange: () => name => name
     }
   }),
-  themeable((customizations, props) => ({
-    customize: {
-      ...props.customize,
-      backgroundColor: customizations.color_checkbox,
-      bulletColor: customizations.color_checkbox_checkmark,
-      borderRadius: customizations.radius_border,
-      textPrimaryColor: customizations.color_text,
-      textSecondaryColor: customizations.color_text_secondary
-    }
-  })),
   uniqueName
 )(Radio)
