+++ Showroom/examples/components.js
@@ -20,7 +20,6 @@ export MouseflowExclude from '../../MouseflowExclude/example'
 export Preview from '../../Preview/example'
 export ProgressBar from '../../ProgressBar/example'
 export Radio from '../../Radio/example'
-export Radio2 from '../../Radio2/example'
 export RadioMark from '../../RadioMark/example'
 export Selector from '../../Selector/example'
 export Strong from '../../Strong/example'
